Add Jaeger collector exporter
